Moved local-storage into shared folder
