+++ src/components/Nutrition/nutrition-results.jsx
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 import {clearAuth} from '../../store/actions/auth';
-import {clearAuthToken} from '../../local-storage';
+import {clearAuthToken} from '../../shared/local-storage';
 import NutritionSearchPage from '../Nutrition/nutrition-search-page';
 import NutritionResultsTotals from '../Nutrition/nutrition-results-totals';
 import ExerciseSearchPage from '../Exercise/exercise-search-page';
