+++ src/shared/local-storage.js
